Code refactor to move functions to more descriptive files
