+++ debunkbot/management/commands/send_replies.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand
 
-from debunkbot.twitter.process_stream import process_stream
+from debunkbot.twitter.process_tweet import process_tweet
 
 
 class Command(BaseCommand):
@@ -8,5 +8,5 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         self.stdout.write(self.style.SUCCESS("Sending reply to selected tweet."))
-        process_stream()
+        process_tweet()
         self.stdout.write(self.style.SUCCESS("Done sending reply"))
